Add files via upload
Cappsule API Fetching with ReactJs with Vite and Designed with TailwindCSS, which fetches Tablets, Tablet Strength, Tablet Forms. And the Prices of the Tablets and the Prices is shown in Ascending order and the Tablet is not available then it will shown as " No Store Is Selling This Product Near You. ". Responsive is only for PC.